Move total value at risk calculation to erc20ValueAtRisk
