+++ src/features/marketplace/pages/essential-dapp/revoke/lib/approvalKeys.ts
@@ -26,22 +26,3 @@ export function filterHiddenBaseAllowances(
 export function getPageBaseAllowances(records: Array<BaseAllowanceType>, page: number): Array<BaseAllowanceType> {
   return records.slice((page - 1) * APPROVALS_PAGE_SIZE, page * APPROVALS_PAGE_SIZE);
 }
-
-export function getTotalValueAtRiskUsd(records: Array<BaseAllowanceType>): number {
-  const maxValues: Record<`0x${ string }`, number> = {};
-
-  records.forEach((item) => {
-    const { address, valueAtRiskUsd } = item;
-
-    if (!valueAtRiskUsd) return;
-
-    if (
-      maxValues[address] === undefined ||
-      valueAtRiskUsd > maxValues[address]
-    ) {
-      maxValues[address] = valueAtRiskUsd;
-    }
-  });
-
-  return Object.values(maxValues).reduce((sum, val) => sum + val, 0);
-}
